Day 9 part 2

+++ src/main/java/uk/co/stevebosman/aoc25/day09/Point.java
@@ -2,10 +2,10 @@ package uk.co.stevebosman.aoc25.day09;
 
 import java.util.List;
 
-public record Point(long x, long y) {
+public record Point(int x, int y) {
   public static Point of(final String x) {
     final String[] positions = x.split(",");
-    return new Point(Long.parseLong(positions[0]), Long.parseLong(positions[1]));
+    return new Point(Integer.parseInt(positions[0]), Integer.parseInt(positions[1]));
   }
 
   public static List<Point> of(final List<String> x) {
@@ -15,6 +15,6 @@ public record Point(long x, long y) {
   }
 
   public static long areaOf(final Point p1, final Point p2) {
-    return (1 + Math.abs(p1.x - p2.x)) * (1 + Math.abs(p1.y - p2.y));
+    return (1L + Math.abs(p1.x - p2.x)) * (1L + Math.abs(p1.y - p2.y));
   }
 }
